Extract BatchGenerator into a top-level class.
